Compile warnings resolved

+++ DuoComparison/src/CSplitOutputProvider.cpp
@@ -194,7 +194,7 @@ void CSplitOutputProvider::FillHeader(CVcfWriter *a_pWriter, bool a_bIsBaseSide)
         a_pWriter->AddHeaderLine("##FILTER=<ID=" + filterNames[k] + ",Description=" + filterDescriptions[k] + ">");
     
     //ADD CONTIG IDs
-    for(int k = 0; k < m_contigs.size(); k++)
+    for(int k = 0; k < (int)m_contigs.size(); k++)
         a_pWriter->AddHeaderLine("##contig=<ID=" + m_contigs[k].name + ",length=" + std::to_string(m_contigs[k].length) + ">");
 
     //ADD REQUIRED SAMPLES
